Revert "Bruk tidslinjer i mapTilOpphørsperioder (#3470)"
    This reverts commit cf3f3e0e9f8c6cc49b340dada466efe2a5e0ca96.
